Naturally propagating moves
Gets rid of all the complicated reserving logic.
Instead, all moves are evaluated as they propagate.

KillSiblings takes a callback that is called on every cube destroyed
and when the move has finished. This allows for doing special effects
during the move, as well as adding up the points an assigning bonues
once the move finishes.