Fix overzealous GPG prompts
The SocialManager could appear multiple times in a scene, causing login
prompts to show up every time the player visited the menu.

In addition, it would keep trying even if the user refused to login.

This commit fixes the duplication and sets the option to not try to
automatically login if the initial attempt fails.

Fixes issue #60