Reduce false positives on the tutorial
Before players would accidentally trigger the fail condition. Now the
condition needs to be triggered 5 times (without any sort of feedback)
before the shortcut kicks in.

There was another issue with the tutorial already. Because of the rule
limiting flooding on the same color, it was impossible to trigger the
shortcut on the kill step.

The flooding rule logic was taken out of PlayerActions and put into
GridLogic, where it has access to the rest of the tutorial variables. This
restriction is just disabled in the tutorial.

Done by using 2 more flags: canFlood and canDestroy. When players perform
the wrong move, the flags are checked and if they did the wrong thing,
no filling or killing happens. In additon, a counter is incremented.

When the counter hits 5, the shortcut activates.

Fixes issue #58